pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "PINN for room heating simulation"
readme = "README.md"
authors = [
    { name = "Elling Svee", email = "86355496+ellingsvee@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jax>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pyyaml>=6.0",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "ty>=0.0.13",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
venvPath = "."
venv = ".venv"
